requirements.txt
flask
flask_sqlalchemy
PyJWT